apps/server/pyproject.toml
[project]
name = "cocktail-server"
version = "0.0.0"
description = "Cocktail の FastAPI サーバ（Gemma + Anima）"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.9",
    "pydantic-settings>=2.5",
    "pillow>=10.4",
    "torch>=2.8",
    "transformers>=4.51",
    "accelerate>=1.0",
    "bitsandbytes>=0.43",
    "diffusers>=0.36",
    # hdae/diffusers-anima は個人リポのため破壊的変更回避のためコミット固定
    "diffusers-anima @ git+https://github.com/hdae/diffusers-anima.git@a5d015b94a368b95e98e6c2aed56c432db24e98c",
    "httpx>=0.27",
    "sentencepiece>=0.2",
    "protobuf>=5",
    "python-multipart>=0.0.18",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cocktail_server"]

[tool.hatch.metadata]
# diffusers-anima を git URL で取るため direct reference を許可
allow-direct-references = true

[tool.uv.sources]
# PyTorch は CUDA 対応ビルドを使う（Blackwell RTX 50xx は CUDA 12.8+ 必須）
torch = { index = "pytorch-cu128" }

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
